+++ lib/modules/theme/theme_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: theme
-// Updated at: 2026-05-27 20:55:18
-// Commit: feat(theme): add async theme stream provider
+// Updated at: 2026-05-28 09:00:38
+// Commit: feat(theme): integrate custom telemetry logger in theme
 
 class themeComponentState {
-  final String commitRef = '1529';
-  final String lastMessage = 'feat(theme): add async theme stream provider';
-  final DateTime timestamp = DateTime.parse('2026-05-27T20:55:18+05:00');
+  final String commitRef = '1539';
+  final String lastMessage = 'feat(theme): integrate custom telemetry logger in theme';
+  final DateTime timestamp = DateTime.parse('2026-05-28T09:00:38+05:00');
 }